requirements.txt
# WebSocket Printer Client Dependencies

# Environment configuration
python-dotenv

# WebSocket client library
python-socketio[asyncio]==5.11.0
aiohttp

# Serial port communication
pyserial==3.5

# USB communication for direct printer connection
pyusb

# PDF generation for Windows printing
reportlab

# Additional utilities
asyncio-mqtt==0.16.1

# Diagnostic tools (optional)
psutil  # For process detection and port diagnostics

# Development dependencies (optional)
pytest==7.4.4
pytest-asyncio==0.23.2
